CoreValidation:  -removed warning for Cortex-A devices.  - reworked test settings.

+++ CMSIS/Core_A/Source/irq_ctrl_gic.c
@@ -184,7 +184,7 @@ __WEAK int32_t IRQ_SetMode (IRQn_ID_t irqn, uint32_t mode) {
     if (val == IRQ_MODE_CPU_ALL) {
       cpu = 0xFFU;
     } else {
-      cpu = val >> IRQ_MODE_CPU_Pos;
+      cpu = (uint8_t)(val >> IRQ_MODE_CPU_Pos);
     }
 
     // Apply configuration if no mode error
